refactor: 히트맵 캘린더 데이터 페칭 로직 다시 컴포넌트로 병합
